+++ libs/sdk-core/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.5 (2022-10-25)
+
+- New method `getProcessingStats` allowing to get stats for resources pending to be processed
+- New enum `StatsRange` to be used with `/processing/stats` endpoint
+
 # 1.0.4 (2022-10-21)
 
 - Update `StatsType` enumeration to be aligned with what's supported by the API